requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
pytest==7.4.3
pytest-flask==1.3.0
qrcode[pil]==7.4.2
requests==2.31.0
beautifulsoup4==4.12.2
Faker==37.12.0
APScheduler==3.11.2
pandas==2.1.3
openpyxl==3.1.2
